add filter flag
this flag adds the ability to be able to filter the results with a regex
so it is possible to get a refined list of one or more interfaces.

For instance, my wifi interface is forgettably called wlp1s0, with the
filter flag I can get it's specific info with `trawl -f w`.